Incubator.TransitionView - remove (move hook to Incubator.Dialog)
