--- demo/src/screens/incubatorScreens/TransitionViewScreen.tsx
@@ -1,54 +1,0 @@
-import React, {Component} from 'react';
-import {View, Button, Incubator} from 'react-native-ui-lib';
-const {TransitionView} = Incubator;
-// @ts-ignore
-import {renderRadioGroup} from '../ExampleScreenPresenter';
-
-interface State {
-  enterDirection: Incubator.TransitionViewDirection;
-  exitDirection: Incubator.TransitionViewDirection;
-  key: number;
-}
-
-export default class TransitionViewScreen extends Component<{}, State> {
-  private ref = React.createRef<typeof TransitionView>();
-  state = {
-    enterDirection: Incubator.TransitionViewDirectionEnum.LEFT,
-    exitDirection: Incubator.TransitionViewDirectionEnum.DOWN,
-    key: 1
-  };
-
-  onPress = () => {
-    this.ref.current?.animateOut();
-  };
-
-  // onAnimationEnd = (type: Incubator.TransitionViewAnimationType) => {
-  //   console.warn('Animation complete', type);
-  // };
-
-  render() {
-    const {key, enterDirection, exitDirection} = this.state;
-    return (
-      <View padding-20 bg-grey80 flex>
-        {renderRadioGroup.call(this, 'Enter direction', 'enterDirection', Incubator.TransitionViewDirectionEnum, {
-          isRow: true
-        })}
-        {renderRadioGroup.call(this, 'Exit direction', 'exitDirection', Incubator.TransitionViewDirectionEnum, {
-          isRow: true
-        })}
-        <Button label="Refresh" onPress={() => this.setState({key: key + 1})}/>
-        <View flex center>
-          <TransitionView
-            key={`${key}`}
-            ref={this.ref}
-            enterFrom={enterDirection}
-            exitTo={exitDirection}
-            // onAnimationEnd={this.onAnimationEnd}
-          >
-            <Button label="Press to remove" onPress={this.onPress}/>
-          </TransitionView>
-        </View>
-      </View>
-    );
-  }
-}
