+++ demo/src/screens/incubatorScreens/index.js
@@ -9,6 +9,5 @@ export function registerScreens(registrar) {
   registrar('unicorn.components.IncubatorTextFieldScreen', () => require('./IncubatorTextFieldScreen').default);
   registrar('unicorn.components.IncubatorToastScreen', () => require('./IncubatorToastScreen').default);
   registrar('unicorn.incubator.PanViewScreen', () => require('./PanViewScreen').default);
-  registrar('unicorn.incubator.TransitionViewScreen', () => require('./TransitionViewScreen').default);
   registrar('unicorn.incubator.WheelPickerScreen', () => gestureHandlerRootHOC(require('./WheelPickerScreen').default));
 }
